HDDS-8987. Address review comments
